src/app/dialog/dialog.component.ts
import { Component, OnInit, Inject, EventEmitter } from '@angular/core';
import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
import { DialogData } from '../models/dialogdata.model';


@Component({
  selector: 'app-dialog',
  templateUrl: './dialog.component.html',
  styleUrls: ['./dialog.component.scss']
})
export class DialogComponent implements OnInit {

  onButton1Click = new EventEmitter();
  onButton2Click = new EventEmitter();

  constructor(
    public dialogRef: MatDialogRef<DialogComponent>,
    @Inject(MAT_DIALOG_DATA) public data: DialogData
  ) { }

  ngOnInit() {
  }

  button1Click() {
    this.dialogRef.close();
    this.onButton1Click.emit();
  }

  button2Click() {
    this.dialogRef.close();
    this.onButton2Click.emit();
  }

}
